requirements.txt
# Core dependencies
azure-storage-blob>=12.19.0
azure-identity>=1.15.0
PySide6>=6.6.1
python-dotenv>=1.0.0

# Testing dependencies
pytest>=7.4.3
pytest-cov>=4.1.0
pytest-qt>=4.2.0
pytest-mock>=3.12.0
coverage>=7.3.2

# Development dependencies
black>=23.7.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.5.1
pre-commit>=3.5.0

psycopg2-binary>=2.9.9
pyinstaller>=6.0.0
pytest-qt==4.4.0
pyyaml>=6.0.0
azure-mgmt-web>=2.0.0
requests>=2.31.0

# Linting
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.1
isort>=5.12.0

# Type checking
types-setuptools>=68.0.0.0
types-requests>=2.31.0.0

# Documentation
sphinx>=7.1.2
sphinx-rtd-theme>=1.3.0

# Build tools
pyinstaller>=6.0.0
twine>=4.0.2
build>=1.0.3

# Development tools
pre-commit>=3.3.3
